docker/requirements.txt
requests
kafka-python
config
websockets